Create Project MedTourEasy
In this project, we will explore the phenomenon using age distribution data to see if you can reproduce a difference in average is at that purely from the changing rates of the left handers over time, refuting the claim of early death for left handers. This notebook uses pandas and Bayesian statistics to analyse the probability of being a certain age at death given that you are reported as  left-handed or right-handed.